refactor model package; update pre-commit;
