+++ pyproject.toml
@@ -28,12 +28,9 @@ store_hazard_v3 = 'scripts.store_hazard_v3:main'
 get_hazard = 'scripts.get_hazard:main'
 query_meta = 'scripts.query_meta:main'
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^3.0.4"
 
-# Newer versions
-#flake8 = "^4.0.1"
-#pytest = "^7.1.2"
-#tox = "^3.25.0"
 
 black  = { version = "^22.3"}
 isort  = { version = "^5.8.0"}
@@ -59,7 +56,6 @@ mkdocs-pymdownx-material-extras = "^2.0.3"
 
 twine  = { version = "^3.3.0"}
 
-pre-commit = {version = "^2.12.0", optional = true}
 toml = {version = "^0.10.2", optional = true}
 bump2version = {version = "^1.0.1"}
 jinja2 = {version = "3.0.3"}
@@ -97,6 +93,7 @@ doc = [
     "mkdocs-material-extension",
     "mkdocs-autorefs"
     ]
+
 
 [tool.black]
 line-length = 120
