+++ toshi_hazard_store/model/__init__.py
@@ -1,22 +1,23 @@
+from .attributes import IMTValuesAttribute, LevelValuePairAttribute
 from .gridded_hazard import CompressedListAttribute, GriddedHazard
 from .gridded_hazard import drop_tables as drop_gridded
 from .gridded_hazard import migrate as migrate_gridded
-from .openquake_v1_model import LevelValuePairAttribute
-from .openquake_v2_model import IMTValuesAttribute
-from .openquake_v3_model import VS30_KEYLEN, HazardAggregation, OpenquakeRealization, ToshiOpenquakeMeta
-from .openquake_v3_model import drop_tables as drop_tables_v3
-from .openquake_v3_model import migrate as migrate_v3
-from .openquake_v3_model import tables as oqv3_tables
-from .openquake_v3_model import vs30_nloc001_gt_rlz_index
+
+# from .openquake_models import tables as oqv3_tables
+# from .openquake_v2_model import
+from .openquake_models import VS30_KEYLEN, HazardAggregation, OpenquakeRealization, ToshiOpenquakeMeta
+from .openquake_models import drop_tables as drop_openquake
+from .openquake_models import migrate as migrate_openquake
+from .openquake_models import vs30_nloc001_gt_rlz_index
 
 
 def migrate():
     """Create the tables, unless they exist already."""
-    migrate_v3()
+    migrate_openquake()
     migrate_gridded()
 
 
 def drop_tables():
     """Drop em"""
-    drop_tables_v3()
+    drop_openquake()
     drop_gridded()
